reorganize service calls for approved users list. Add types
